[ADD] dir parsing && getpath && utils *.c
