+++ null.d
@@ -1,0 +1,1 @@
+null.o: /dev/null
